feat: rename research note function to document for consistency
- Renamed create-research-note.ts to create-document.ts to better align with function naming conventions
- Updated function key from 'dxos.org/function/research/create-document' to match new name
- Updated imports and references in index.ts and research.test.ts to use new module name
- No changes to function behavior or implementation, purely renaming for clarity
