+++ packages/core/assistant-toolkit/src/functions/research/create-document.ts
