+++ packages/core/assistant-toolkit/src/functions/research/index.ts
@@ -2,7 +2,7 @@
 // Copyright 2025 DXOS.org
 //
 
-import { default as createResearchNote } from './create-research-note';
+import { default as createDocument } from './create-document';
 import { default as research } from './research';
 
 export * from './graph';
@@ -10,5 +10,5 @@ export * from './research-graph';
 export * from './types';
 
 export namespace Research {
-  export const tools = [research, createResearchNote] as const;
+  export const tools = [research, createDocument] as const;
 }
